chore: remove vta-sdk local patch and fix clippy warnings (#278)
Remove vta-sdk from [patch.crates-io] now that it is published on
crates.io. Fix all clippy warnings across the workspace (collapsible
if statements, redundant field names, needless borrows, derivable
impls, unnecessary casts, and type complexity).

+++ crates/messaging/affinidi-messaging-mediator/affinidi-messaging-mediator-processors/src/lib/forwarding/processor.rs
@@ -353,7 +353,7 @@ impl ForwardingProcessor {
     }
 
     fn calculate_backoff(&self, consecutive_failures: u32) -> Duration {
-        let base = self.config.initial_backoff_ms as u64;
+        let base = self.config.initial_backoff_ms;
         let max = self.config.max_backoff_ms;
         let backoff = base.saturating_mul(2u64.saturating_pow(consecutive_failures.min(10)));
         Duration::from_millis(backoff.min(max))
